refactor(app): split demo.py into modules, keep behavior; fix CSS :root selector; add app_factory
